Use FW_SEARCH_PATH to find the HistogramFile.
The fcl parameter HistogramFile can now specify an absolute path
or a relative path.  If an absolute path is used, the file is
looked for in that specific location only.  If a relative path
is used, the file is first looked for using the relative path
applied to the current working directory of the job, and if not
found, then it is applied to each directory in FW_SEARCH_PATH
in order until found.