Fix redmine issue #20554 - AuxDetSensitive with ID 0 has no energy deposit.
The fallback code which creates a sensitive detector for each AuxDet if
there are no AuxDetSensitive in the geometry (supports old gdml files
created before the introduction of AuxDetSensitive) was being invoked
unconditionally on every execution.  This resulted in Geant4 deleting
the sensitive detectors with id 0, so no data in that channel was being
recorded.  The fix is to avoid doing the fallback if there really was
AuxDetSensitive in the geometry.