Removed optimizations which do not universally apply (Redmine issue #24600).
The optimizations are related to the mapping of scintillation visibility
only in the active volume, and on the complete opaqueness of the cathode.
These optimizations are now controlled by flags.
The first optimization is enabled only when there is only one cryostat
and the semi-analytic model is being used.
The other is never enabled, because we currently do not have a way to determine
whether the cathode is fully opaque.