New bool-type variable UserD added
The function of UserD is to determine whether applying user-defined region or not. The default value of UserD is false. Also, default values of all user-defined boundary variables, ex: BminX, BmaxX, are all set as 0; Only when UserD is set as true, can users  set the region of their own interests. When UserD is not mentioned in fhicl script, the default region is just the boundary of whole cryostat.