MCRecoEdep: simplified TPC check, and checking only planes on that TPC
The previous algorithm would test all the wire planes in the detector for each energy deposit.
Plus, the containment check was buggy (`NearestWireID()` behaviour is not reliable for points outside the plane)
and it ended up including planes that should not have been.
Also, the containment check is now quiet and does not trigger any exception.